take two: merge TU19 changes (december 2014)

+++ Minecraft.World/Recipes/ShapedRecipy.cpp
@@ -120,14 +120,11 @@ void ShapedRecipy::collectRequirements(INGREDIENTS_REQUIRED* pIngReq) {
     TempIngReq.iType = ((width > 2) || (height > 2)) ? RECIPE_TYPE_3x3
                                                      : RECIPE_TYPE_2x2;  // 3x3
     // 3x3
-    // 4jcraft, genuinly what is this garbage code
     TempIngReq.uiGridA = new unsigned int[9];
     TempIngReq.iIngIDA = new int[9];
     TempIngReq.iIngValA = new int[9];
     TempIngReq.iIngAuxValA = new int[9];
 
-    // 4jcraft,yes, yes!!
-    // use winapi and inbetween use a cstd function u could have used!
     ZeroMemory(TempIngReq.iIngIDA, sizeof(int) * 9);
     ZeroMemory(TempIngReq.iIngValA, sizeof(int) * 9);
     memset(TempIngReq.iIngAuxValA, Recipes::ANY_AUX_VALUE, sizeof(int) * 9);
@@ -140,10 +137,8 @@ void ShapedRecipy::collectRequirements(INGREDIENTS_REQUIRED* pIngReq) {
 
                 if (expected != NULL) {
                     int iAuxVal = expected->getAuxValue();
-                    // 4jcraft, added cast to uint (shift of negativ num,
-                    // undefined)
-                    TempIngReq.uiGridA[x + y * 3] =
-                        expected->id | (unsigned int)iAuxVal << 24;
+                    TempIngReq.uiGridA[x + y * 3] = expected->id | iAuxVal
+                                                                       << 24;
 
                     bFound = false;
                     for (j = 0; j < TempIngReq.iIngC; j++) {
